# swagger-new.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IronLogic Fitness Training API",
    "description": "API for managing fitness training data including clients, activities, schedules, and benchmarks",
    "version": "1.0.0",
    "contact": {
      "name": "IronLogic API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/lift-benchmarks": {
      "get": {
        "tags": ["Benchmarks"],
        "summary": "Get all lift benchmarks",
        "description": "Retrieve a list of all lift benchmarks",
        "responses": {
          "200": {
            "description": "List of lift benchmarks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LiftBenchmark"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "post": {
        "tags": ["Benchmarks"],
        "summary": "Create a new lift benchmark",
        "description": "Create a new lift benchmark",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "weight"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "description": "Benchmark name"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "Benchmark notes"
                  },
                  "weight": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Weight achieved"
                  },
                  "benchmarkTemplateId": {
                    "type": "string",
                    "description": "Reference to benchmark template"
                  }
                }
              },
              "example": {
                "name": "Bench Press Max",
                "notes": "1RM bench press",
                "weight": 225,
                "benchmarkTemplateId": "template123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lift benchmark created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/LiftBenchmark"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
    "/lift-benchmarks/{id}": {
      "get": {
        "tags": ["Benchmarks"],
        "summary": "Get lift benchmark by ID",
        "description": "Get a specific lift benchmark by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Lift benchmark ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Lift benchmark found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/LiftBenchmark"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Lift benchmark not found"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "put": {
        "tags": ["Benchmarks"],
        "summary": "Update lift benchmark",
        "description": "Update a lift benchmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Lift benchmark ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "weight"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "weight": {
                    "type": "number",
                    "minimum": 0
                  },
                  "benchmarkTemplateId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lift benchmark updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/LiftBenchmark"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Lift benchmark not found"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "delete": {
        "tags": ["Benchmarks"],
        "summary": "Delete lift benchmark",
        "description": "Delete a lift benchmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Lift benchmark ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Lift benchmark deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Lift benchmark not found"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LiftBenchmark": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Benchmark ID"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Benchmark name"
          },
          "notes": {
            "type": "string",
            "maxLength": 1000,
            "description": "Benchmark notes"
          },
          "weight": {
            "type": "number",
            "minimum": 0,
            "description": "Weight achieved"
          },
          "benchmarkTemplateId": {
            "type": "string",
            "description": "Reference to benchmark template"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": ["name", "weight"]
      },
...
    }
  }
} 